+++ src/Draco.Compiler/Internal/Codegen/DracoIrCodegen.cs
@@ -9,7 +9,6 @@ using Draco.Compiler.Internal.Symbols;
 using Draco.Compiler.Internal.Symbols.Source;
 using Draco.Compiler.Internal.Symbols.Synthetized;
 using Draco.Compiler.Internal.Types;
-using static System.Net.Mime.MediaTypeNames;
 using IrType = Draco.Compiler.Internal.DracoIr.Type;
 
 namespace Draco.Compiler.Internal.Codegen;
